Unit tests for unmarshal_request(..)
